PR feedback fixes

+++ src/devices/LidarLiteV3/samples/LidarLiteV3.Sample.cs
@@ -16,8 +16,8 @@ namespace Iot.Device.TimeOfFlight.LidarLiteV3Samples
             using (var llv3 = new LidarLiteV3(CreateI2cDevice()))
             {
                 // Take 10 measurements, each one second apart.
-                for(int i = 0; i < 10; i++) {
-                    ushort currentDistance = llv3.MeasureDistance();
+                for (int i = 0; i < 10; i++) {
+                    int currentDistance = llv3.MeasureDistance();
                     Console.WriteLine($"Current Distance: {currentDistance} cm");
                     Thread.Sleep(1000);
                 }
